Improvements from code review
