+++ jest.config.puppeteer.js
@@ -151,7 +151,7 @@ module.exports = {
 
 	// An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
 	testPathIgnorePatterns: [
-		"/node_modules/", "/jest/testUtils"
+		"/node_modules/", "/puppeteer/testUtils"
 	],
 
 	// The regexp pattern or array of patterns that Jest uses to detect test files
