+++ package.json
@@ -14,7 +14,7 @@
     "test-nw:local": "nightwatch -c ./test/nightwatch/config/nightwatch.conf.local.js",
     "test-nw:basic": "npm run nightwatch-bs -- --group basic --env chrome,firefox,safari,edge,galaxy_s8",
     "test-nw:extended": "npm run nightwatch-bs -- --group extended --env chrome",
-    "test-e2e": "jest -c jest.config.puppeteer.js --detectOpenHandles",
+    "test-e2e": "jest -c jest.config.puppeteer.js",
     "test-e2e-local": "start-server-and-test demo-server http://localhost:3002 test-e2e",
     "jest": "jest",
     "coverage": "export COVERAGE=true && karma start && unset COVERAGE",
@@ -65,7 +65,6 @@
     "karma-webpack": "^3.0.5",
     "nightwatch": "git@github.com:Financial-Times/nightwatch.git",
     "origami-build-tools": "^8.0.0",
-    "parse-url": "^5.0.1",
     "pre-git": "^3.17.1",
     "puppeteer": "^1.19.0",
     "qunitjs": "^1.20.0",
